feat(my-programs): split Attendance to its own path, per-tab subtitles
Move the staff My Programs attendance inbox from /my-programs to its own
subtab path /my-programs/attendance; bare /my-programs now redirects there.
Drop the section title and shared subtitle; each subtab gets its own subtitle.

Registry: point the staff entry at the new path, exclude the redirect-only
/my-programs (matches the /attendance precedent) so the drift guard stays green.
